shots type modify int

+++ fusion_stat/models/base.py
@@ -12,5 +12,5 @@ class ParamsDict(typing.TypedDict):
 
 
 class FBrefShootingDict(typing.TypedDict):
-    shots: float
+    shots: int
     xg: float
